Configurations cleanup and refactoring (#116)
* Athena - clean up and refactoring
