+++ properties/test/application.properties
@@ -25,12 +25,6 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.show-sql=true
 
-arachne.datasource.url=jdbc:postgresql://localhost:5432/arachne_portal
-arachne.datasource.driver-class-name=org.postgresql.Driver
-arachne.datasource.username=ohdsi
-arachne.datasource.password=
-arachne.jpa.hibernate.ddl-auto=none
-
 athena.v4.datasource.url=jdbc:postgresql://localhost:5432/athena_cdm_v4_5
 athena.v4.datasource.driver-class-name=org.postgresql.Driver
 athena.v4.datasource.username=ohdsi
