+++ src/main/java/com/odysseusinc/athena/repositories/v5/ConceptV5Repository.java
@@ -24,14 +24,12 @@ package com.odysseusinc.athena.repositories.v5;
 
 import com.odysseusinc.athena.model.athenav5.ConceptV5;
 import java.util.List;
-import javax.persistence.PersistenceContext;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-@PersistenceContext(unitName = "athenaEntityManagerFactory")
 public interface ConceptV5Repository extends JpaRepository<ConceptV5, Long> {
 
     @Query(nativeQuery = true,
