+++ src/main/java/com/odysseusinc/athena/service/impl/ConceptServiceImpl.java
@@ -22,19 +22,9 @@
 
 package com.odysseusinc.athena.service.impl;
 
-import static java.util.stream.Collectors.toList;
-import static org.apache.solr.common.params.CursorMarkParams.CURSOR_MARK_PARAM;
-import static org.apache.solr.common.params.CursorMarkParams.CURSOR_MARK_START;
-
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
-import com.odysseusinc.athena.api.v1.controller.converter.ConceptSearchDTOToSolrQuery;
-import com.odysseusinc.athena.api.v1.controller.converter.ConceptSearchResultToDTO;
-import com.odysseusinc.athena.api.v1.controller.converter.SolrDocumentToConceptDTO;
-import com.odysseusinc.athena.api.v1.controller.dto.ConceptDTO;
-import com.odysseusinc.athena.api.v1.controller.dto.ConceptSearchDTO;
-import com.odysseusinc.athena.api.v1.controller.dto.ConceptSearchResultDTO;
 import com.odysseusinc.athena.model.athenav5.ConceptAncestor;
 import com.odysseusinc.athena.model.athenav5.ConceptAncestorRelationV5;
 import com.odysseusinc.athena.model.athenav5.ConceptAncestor_;
@@ -50,35 +40,9 @@ import com.odysseusinc.athena.repositories.v5.ConceptRelationshipV5Repository;
 import com.odysseusinc.athena.repositories.v5.ConceptV5Repository;
 import com.odysseusinc.athena.repositories.v5.RelationshipV5Repository;
 import com.odysseusinc.athena.service.ConceptService;
-import com.odysseusinc.athena.service.SolrService;
 import com.odysseusinc.athena.service.VocabularyConversionService;
 import com.odysseusinc.athena.service.aspect.LicenseCheck;
 import com.odysseusinc.athena.service.graph.RelationGraphParameter;
-import com.odysseusinc.athena.service.impl.solr.SearchResult;
-import com.odysseusinc.athena.service.writer.FileHelper;
-import com.odysseusinc.athena.util.extractor.ConceptFieldsExtractor;
-import com.opencsv.CSVWriter;
-import java.io.File;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-import javax.persistence.EntityManager;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Path;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
-import javax.validation.constraints.NotNull;
-import org.apache.solr.client.solrj.SolrQuery;
-import org.apache.solr.client.solrj.SolrServerException;
-import org.apache.solr.client.solrj.response.QueryResponse;
-import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -86,17 +50,25 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityManager;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Path;
+import javax.persistence.criteria.Predicate;
+import javax.persistence.criteria.Root;
+import javax.validation.constraints.NotNull;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+
 @Service
-@Transactional(readOnly = true)
+@Transactional(readOnly = true, transactionManager = "athenaV5TransactionManager")
 public class ConceptServiceImpl implements ConceptService {
+
     @Autowired
     private ConceptV5Repository conceptRepository;
-    @Autowired
-    private SolrService solrService;
-    @Autowired
-    private ConceptSearchDTOToSolrQuery converterToSolrQuery;
-    @Autowired
-    private ConceptSearchResultToDTO converter;
     @Autowired
     private ConceptAncestorRelationV5Repository ancestorRelationV5Repository;
     @Autowired
@@ -104,16 +76,12 @@ public class ConceptServiceImpl implements ConceptService {
     @Autowired
     private RelationshipV5Repository relationshipV5Repository;
     @Autowired
-    private FileHelper fileHelper;
-    @Autowired
     private VocabularyConversionService conversionService;
     @Autowired
     private UserService userService;
     @Autowired
     @Qualifier("athenaV5EntityManagerFactory")
     private EntityManager entityManager;
-    @Value("${csv.separator:;}")
-    private Character separator;
 
     @Value("${csv.file.name.searched.concepts}")
     private String csvFileName;
@@ -240,60 +208,6 @@ public class ConceptServiceImpl implements ConceptService {
     public List<RelationshipV5> getAllRelationships(Long id) {
 
         return relationshipV5Repository.findRelationships(id);
-    }
-
-    @Override
-    public ConceptSearchResultDTO search(ConceptSearchDTO searchDTO) throws IOException, SolrServerException {
-
-        List<String> v5Ids = conversionService.getUnavailableVocabularies();
-        SolrQuery solrQuery = converterToSolrQuery.createQuery(searchDTO, v5Ids);
-        QueryResponse solrResponse = solrService.search(solrQuery);
-        List<SolrDocument> solrDocumentList = solrResponse.getResults();
-        return converter.convert(new SearchResult<>(solrQuery, solrResponse, solrDocumentList), v5Ids);
-    }
-
-    @Override
-    public void generateCSV(ConceptSearchDTO searchDTO, OutputStream osw) throws IOException, SolrServerException {
-
-        SolrQuery solrQuery = converterToSolrQuery.convertForCursor(searchDTO);
-        String cursorMark = CURSOR_MARK_START;
-        boolean done = false;
-        boolean first = true;
-
-        String name = fileHelper.getTempPath(UUID.randomUUID().toString());
-        File temp = new File(name);
-        while (!done) {
-
-            try (CSVWriter csvWriter = new AthenaCSVWriter(name, separator)) {
-                if (first) {
-                    csvWriter.writeNext(new String[]{"Id", "Code", "Name", "Concept Class Id", "Domain", "Vocabulary",
-                            "Invalid Reason", "Standard Concept"}, false);
-                    first = false;
-                }
-                solrQuery.set(CURSOR_MARK_PARAM, cursorMark);
-                QueryResponse solrResponse = solrService.search(solrQuery);
-                List<SolrDocument> solrDocuments = solrResponse.getResults();
-                List<ConceptDTO> concepts = solrDocuments.stream()
-                        .map(SolrDocumentToConceptDTO::convert)
-                        .collect(toList());
-                writeAll(csvWriter, concepts);
-                String nextCursorMark = solrResponse.getNextCursorMark();
-                if (cursorMark.equals(nextCursorMark)) {
-                    done = true;
-                }
-                cursorMark = nextCursorMark;
-                csvWriter.flush(true);
-            } finally {
-                Files.copy(temp.toPath(), osw);
-                temp.delete();
-            }
-        }
-    }
-
-    private void writeAll(CSVWriter csvWriter, List<ConceptDTO> concepts) throws IOException {
-
-        ConceptFieldsExtractor extractor = new ConceptFieldsExtractor();
-        csvWriter.writeAll(new ArrayList<>(extractor.extractForAll(concepts)));
     }
 
 }
